requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-filter==2.0.0
django-phonenumber-field==2.1.0
django-widget-tweaks==1.4.3
idna==2.8
oauthlib==2.1.0
olefile==0.46
phonenumbers==8.10.2
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
urllib3==1.24.1
